propbot/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pydantic==2.10.3
python-dotenv==1.0.1
openai==1.58.1
numpy==2.2.0
